+++ client/ayon_core/tools/publisher/models/__init__.py
@@ -1,6 +1,9 @@
+from .create import CreatorItem
 from .publish import PublishReportMaker
 
 
 __all__ = (
+    "CreatorItem",
+
     "PublishReportMaker",
 )
